+++ lib/canvas-workflow.rb
@@ -1,0 +1,6 @@
+require 'canvas/workflow'
+
+module Canvas
+  module Workflow
+  end
+end