+++ lib/canvas/workflow/cli.rb
@@ -1,0 +1,11 @@
+require 'canvas/workflow/cli/builder'
+require 'canvas/workflow/cli/deployer'
+require 'canvas/workflow/cli/main'
+require 'canvas/workflow/cli/pusher'
+
+module Canvas
+  module Workflow
+    module CLI
+    end
+  end
+end