+++ lib/canvas/workflow/jekyll.rb
@@ -1,0 +1,13 @@
+require 'jekyll' # jekyll & liquid
+
+require 'canvas/workflow/jekyll/assignment'
+require 'canvas/workflow/jekyll/file'
+require 'canvas/workflow/jekyll/gist'
+require 'canvas/workflow/jekyll/icon'
+
+module Canvas
+  module Workflow
+    module Jekyll
+    end
+  end
+end